requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
jinja2==3.1.2
python-docx==1.1.0
psycopg[binary]==3.1.18
redis==5.0.1
celery==5.3.6
asyncssh==2.14.2
cryptography==42.0.8
gunicorn==21.2.0
alembic==1.13.1
croniter==2.0.5
